Move Radio Button tests
